初步完成厂房管理模块的初始化

+++ src/main/java/edu/cqupt/mislab/erp/game/compete/operation/produce/model/prodline/entity/ProdlineBasicInfo.java
@@ -69,7 +69,7 @@ public class ProdlineBasicInfo implements Serializable {
 
     @Column(nullable = false, updatable = false)
     @Comment(comment = "该生产线对产品生产周期的影响情况")
-    private Integer extraPeriod;
+    private Double extraPeriod;
 
     @Basic(optional = false)
     @Comment(comment = "该条设置是否启用")
